Command processors now in separated files, it should be easier to test. Fixed bot crash when message entities is null.
