+++ bot/src/main/java/ru/tinkoff/edu/java/bot/tg/command/CommandProcessor.java
@@ -1,4 +1,4 @@
-package ru.tinkoff.edu.java.bot.tg;
+package ru.tinkoff.edu.java.bot.tg.command;
 
 import com.pengrad.telegrambot.model.Message;
 import com.pengrad.telegrambot.request.SendMessage;
